requirements.txt
# Core ML and AI packages
streamlit==1.28.1
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.8.0

faiss-cpu==1.7.4

# Data processing
pandas==2.1.3
numpy==1.24.3
nltk==3.8.1

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0

# Additional utilities
requests==2.31.0
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0

# Optional: For better performance
accelerate==0.24.1
